pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whatsupdoc"
version = "0.1.0"
description = "A Slack RAG chatbot using Vertex AI RAG Engine and Gemini"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "whatsupdoc", email = "noreply@example.com"}
]
dependencies = [
    "python-dotenv>=1.0.1",
    "google-auth>=2.40.3",
    "google-cloud-aiplatform>=1.95.1",
    "google-genai>=1.29.0",
    "slack-bolt>=1.20.1",
    "requests>=2.32.0",
    "structlog>=24.1.0",
    "flask>=3.1.1",
    "pydantic>=2.9.0",
    "tenacity>=9.0.0",
]

[project.scripts]
whatsupdoc = "whatsupdoc.app:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "requests-mock>=1.12.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-requests>=2.32.0",
]

[project.optional-dependencies]
cloud-functions = [
    "functions-framework>=3.5.0",
    "google-cloud-aiplatform>=1.108.0",
    "cloudevents>=1.11.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings"
]
markers = [
    "integration: marks tests as integration tests (may be slow)",
    "unit: marks tests as unit tests (fast)",
    "requires_gcp: marks tests that require GCP credentials",
    "requires_slack: marks tests that require Slack tokens"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "C", "I", "N", "D", "UP", "ANN"]
ignore = ["D100", "D104"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
